Edit Goals Page Cleanup
Move majority of page content to new "EditGoalsForm" component.
Added admin protection on page.
